pyproject.toml
[build-system]
requires = ["hatchling", "setuptools"]
build-backend = "hatchling.build"

[project]
name = "setcover-gnn"
version = "0.1.0"
description = "GNN-based solver for Set Cover problems using PyTorch Lightning"
authors = [
    {name = "HrpperCroFT", email = "hrppercroft@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "lightning==2.6.0",
    "networkx==3.6.1",
    "numpy==1.26.4",
    "qubovert==1.2.5",
    "sympy==1.14.0",
    "scipy==1.16.3",
    "click",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
    "pytest-mock",
    "black",
    "isort",
    "flake8",
    "mypy",
    "pre-commit",
    "jupyter",
    "ipython",
]
test = [
    "hydra-core",
    "omegaconf",
    "mlflow",
]
plotting = [
    "matplotlib",
    "pandas",
]

[tool.hatch.build.targets.wheel]
packages = ["src/setcover_gnn"]